removed canvas.js and added canvas_controller instead. Updated script source on classroom show page. Added transform to canvas to make desks visible. Desks are auto-generated but blurry (likely because the canvas is actually a lot bigger than what is shown?). Will continue to look into it.
