+++ app/views/classrooms/show.html.erb
@@ -24,7 +24,7 @@
     <h3><% @classroom.name %></h3>
     <h6 class="my-1">Back</h6>
     <canvas></canvas>
-    <script src="../../javascript/canvas.js"></script>
+    <script src="app/javascript/canvas.js"></script>
     <h6 class="my-1">Front</h6>
   </div>
 </div>
